chore: move config package to internal
