+++ cli/cmd/run.go
@@ -9,7 +9,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/elastic/e2e-testing/cli/config"
+	"github.com/elastic/e2e-testing/internal/config"
 	"github.com/elastic/e2e-testing/internal/deploy"
 	log "github.com/sirupsen/logrus"
