+++ cli/main.go
@@ -6,7 +6,7 @@ package main
 
 import (
 	"github.com/elastic/e2e-testing/cli/cmd"
-	"github.com/elastic/e2e-testing/cli/config"
+	"github.com/elastic/e2e-testing/internal/config"
 )
 
 func init() {
