+++ e2e/_suites/kubernetes-autodiscover/autodiscover_test.go
@@ -26,8 +26,8 @@ import (
 	flag "github.com/spf13/pflag"
 	"go.elastic.co/apm"
 
-	"github.com/elastic/e2e-testing/cli/config"
 	"github.com/elastic/e2e-testing/internal/common"
+	"github.com/elastic/e2e-testing/internal/config"
 	"github.com/elastic/e2e-testing/internal/deploy"
 	"github.com/elastic/e2e-testing/internal/kubernetes"
 	"github.com/elastic/e2e-testing/internal/shell"
