+++ internal/config/compose/profiles/fleet/default/kibana.config.yml
