+++ internal/config/compose/profiles/fleet/elasticsearch-roles.yml
