+++ internal/config/compose/profiles/fleet/preconfigured-policies/kibana.config.yml
