+++ internal/config/compose/services/elastic-agent/centos/docker-compose.yml
