+++ internal/config/compose/services/elastic-agent/cloud/docker-compose.yml
