+++ internal/config/compose/services/elastic-agent/debian/docker-compose.yml
