+++ internal/config/compose/services/elastic-agent/docker-compose.yml
