+++ internal/config/compose/services/elastic-agent/fleet-server/docker-compose.yml
