+++ internal/config/compose/services/elasticsearch/docker-compose.yml
