+++ internal/config/compose/services/kibana/docker-compose.yml
