+++ internal/config/compose/services/metricbeat/docker-compose.yml
