+++ internal/config/compose/services/opbeans-go/docker-compose.yml
