+++ internal/config/compose/services/opbeans-java/docker-compose.yml
