+++ internal/config/compose/services/vsphere/docker-compose.yml
