+++ internal/config/kubernetes/base/elasticsearch/configmap.yaml
