+++ internal/config/kubernetes/base/elasticsearch/deployment.yaml
