+++ internal/config/kubernetes/base/elasticsearch/kustomization.yaml
