+++ internal/config/kubernetes/base/elasticsearch/service.yaml
