+++ internal/config/kubernetes/base/fleet-server/cluster-role.yaml
