+++ internal/config/kubernetes/base/fleet-server/role.yaml
