+++ internal/config/kubernetes/base/kibana/configmap.yaml
