+++ internal/config/kubernetes/base/kibana/deployment.yaml
