+++ internal/config/kubernetes/base/kibana/kustomization.yaml
