+++ internal/config/kubernetes/base/kibana/service.yaml
