+++ internal/config/kubernetes/base/package-registry/deployment.yaml
