+++ internal/config/kubernetes/base/package-registry/kustomization.yaml
