+++ internal/config/kubernetes/base/package-registry/service.yaml
