+++ internal/config/kubernetes/kind.yaml
