+++ internal/deploy/compose.go
@@ -10,7 +10,7 @@ import (
 	"path"
 	"path/filepath"
 
-	"github.com/elastic/e2e-testing/cli/config"
+	"github.com/elastic/e2e-testing/internal/config"
 	"github.com/elastic/e2e-testing/internal/io"
 	state "github.com/elastic/e2e-testing/internal/state"
 	"go.elastic.co/apm"
